requirements.txt
asgiref==3.7.2
cfgv==3.3.1
colorama==0.4.6
distlib==0.3.6
Django==5.0.1
django-bootstrap3==23.6
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
django-debug-toolbar==4.2.0
Faker==19.13.0
filelock==3.12.2
identify==2.5.24
iniconfig==2.0.0
nodeenv==1.8.0
packaging==23.1
pip==23.3.2
platformdirs==3.8.0
pluggy==1.2.0
pre-commit== 3.3.3
py4j==0.10.9.7
pytest==7.4.0
python-dateutil==2.8.2
PyYAML==6.0.1
setuptools==65.5.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.23.1
requests==2.31
Pillow==10.4.0
misaka==2.1.0
psycopg2==2.9.9
